requirements.txt
# ===========================================
# GameTestAgent 依赖清单
# ===========================================

# -------------------------------------------
# 核心依赖
# -------------------------------------------
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0

# -------------------------------------------
# MCP协议支持
# -------------------------------------------
mcp>=1.0.0

# -------------------------------------------
# 大模型API
# -------------------------------------------
openai>=1.0.0
httpx>=0.25.0

# -------------------------------------------
# 文档解析
# -------------------------------------------
python-docx>=0.8.11
PyPDF2>=3.0.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# -------------------------------------------
# 数据处理和表检查
# -------------------------------------------
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.1
numpy>=1.24.0

# -------------------------------------------
# 测试用例导出
# -------------------------------------------
XlsxWriter>=3.1.0

# -------------------------------------------
# XMind支持（可选）
# 安装后可导出真正的.xmind文件
# 否则会回退到JSON/Markdown格式
# -------------------------------------------
# xmind>=1.2.0

# -------------------------------------------
# 数据库支持
# -------------------------------------------
sqlalchemy>=2.0.0
pymysql>=1.1.0
psycopg2-binary>=2.9.0

# -------------------------------------------
# 项目管理平台API
# -------------------------------------------
jira>=3.5.0

# -------------------------------------------
# 工具库
# -------------------------------------------
typing-extensions>=4.8.0
aiohttp>=3.9.0

# -------------------------------------------
# 日志和调试
# -------------------------------------------
loguru>=0.7.0
rich>=13.0.0

# -------------------------------------------
# 测试（开发用）
# -------------------------------------------
pytest>=7.4.0
pytest-asyncio>=0.21.0
